kvartirka/pyproject.toml
[tool.poetry]
name = "kvartirka_test"
version = "0.1.0"
description = "Test app for Kvartirka.ru"
authors = ["Irek Gimatov <gimatov@list.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.6"
djangorestframework = "^3.12.4"
psycopg2-binary = "^2.9.1"
django-treebeard = "^4.5.1"
PyYAML = "^5.4.1"
uritemplate = "^3.0.1"
toml = "^0.10.2"
djangorestframework-simplejwt = "^4.7.2"
coreapi = "^2.3.3"
drf-yasg = {extras = ["validation"], version = "^1.20.0"}
django-environ = "^0.4.5"
gunicorn = "^20.1.0"
celery = {extras = ["redis"], version = "^5.1.2"}
django-redis = "^5.0.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"
django-debug-toolbar = "^3.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
